s3 image posting and sql img url posting complete
