+++ routes/SQLImageRouter.js
@@ -52,25 +52,8 @@ router.get('/:deptID', function(req, res) {
 });
 
 // Post to SQL DB, not to S3
-router.post('/', function (req, res, next) {
-  pool.connect(function (err, client, done) {
-    try {
-      if (err) {
-        res.sendStatus(500);
-      }
-      client.query('INSERT INTO images (img_url, department_id) VALUES [$1, $2])', [req.body.url, req.file.department],
-                  function (err) {
-        if (err) {
-          console.log('Error inserting into db', err);
-          return res.sendStatus(500);
-        }
-        res.sendStatus(200);
-      });
-    } finally {
-      done();
-    }
-  });
-});
+// this takes place in s3ImageRouter.js file as there is no way to predict or
+// control the file name assigned by multer and amazon s3.
 
 //deletes entries from SQL database, not S3
 router.delete('/:id', function (req, res, next) {
